==> 00_EDA.ipynb <==
import matplotlib.pyplot as plt
import plotly.graph_objs as go
from plotly.subplots import make_subplots
# from wordcloud import WordCloud
import numpy as np
import seaborn as sns
%matplotlib inline
from plotly.offline import init_notebook_mode, iplot

import pandas as pd

data = pd.read_csv("amazon_reviews.csv", usecols=[ 'reviewerName', 'overall', 'reviewText', 'reviewTime',
                                                   'day_diff', 'helpful_yes', 'helpful_no', 'total_vote',
                                                   'score_pos_neg_diff', 'score_average_rating',
                                                   'wilson_lower_bound']
                  )

constraints = ['#581845','#C70039','#2E4053','#1ABC9C','#7F8C8D']

def categorical_variable_summary(df, column_name):
    fig = make_subplots(rows=1,cols=2,
                        subplot_titles=('Countplot','Percentages'),
                        specs=[[{"type": "xy"}, {'type':'domain'}]])

    fig.add_trace(go.Bar( y = df[column_name].value_counts().values.tolist(), 
                          x = [str(i) for i in df[column_name].value_counts().index], 
                          text = df[column_name].value_counts().values.tolist(),
                          textfont = dict(size=15),
                          name = column_name,
                          textposition = 'auto',
                          showlegend=False,
                          marker=dict(color = constraints,
                                      line=dict(color='#DBE6EC',
                                                width=1))),
                  row = 1, col = 1)
    
    fig.add_trace(go.Pie(labels= df[column_name].value_counts().keys(),
                         values= df[column_name].value_counts().values,
                         textfont = dict(size = 20),
                         textposition='auto',
                         showlegend = False,
                         name = column_name,
                         marker=dict(colors=constraints)),
                  row = 1, col = 2)
    
    fig.update_layout(title={'text': column_name,
                             'y':0.9,
                             'x':0.5,
                             'xanchor': 'center',
                             'yanchor': 'top'},
                      template='plotly_white')
    
    iplot(fig)

categorical_variable_summary(df=data, column_name="overall")
