==> 01_Data_Split.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
import os

data = pd.read_csv("amazon_reviews.csv", usecols=[ 'reviewerName', 'overall', 'reviewText', 'reviewTime',
                                                   'day_diff', 'helpful_yes', 'helpful_no', 'total_vote',
                                                   'score_pos_neg_diff', 'score_average_rating',
                                                   'wilson_lower_bound']
                  )

data.head()

data.columns

data["overall"].value_counts()

for text in data[data["overall"]==3]["reviewText"].tolist():
    print(text, end = "\n\n")

reviews = data[["reviewText", "overall"]]

reviews = reviews[reviews["overall"]!=4]

reviews["sentiment"] = reviews["overall"].apply(lambda x: 1 if x>3 else 0)

reviews["sentiment"].value_counts()

train_data, test_data = train_test_split(reviews, stratify=reviews["sentiment"],
                                         test_size=0.2, random_state=42)

test_data, validation_data = train_test_split(test_data, stratify=test_data["sentiment"],
                                             test_size=0.5, random_state=42)

train_data, test_data

train_data["sentiment"].value_counts()

test_data["sentiment"].value_counts()

validation_data["sentiment"].value_counts()

os.makedirs("data_v2/", exist_ok=True)

train_data.to_csv("data_v2/train_data.csv", index=False)
test_data.to_csv("data_v2/test_data.csv", index=False)
validation_data.to_csv("data_v2/validation_data.csv", index=False)
