==> 03_DATA_AUGMENTATION_Training_Data_Random_Insertion.ipynb <==
import pandas as pd
from transformers import pipeline
import spacy
import random
nlp = spacy.load("en_core_web_sm")

train_data = pd.read_csv("data_v2/train_data.csv")

train_data["sentiment"].value_counts()

unmasker_1 = pipeline("fill-mask", model="roberta-base")
unmasker_2 = pipeline("fill-mask", model="bert-base-cased")

train_data = train_data[train_data["sentiment"]==0]

train_data.shape

def get_sents(x):
    for sent in nlp(x).sents:
        yield sent.text


def get_random_word_added_1(input_text):
    augmented_texts = []
    
    for sent in get_sents(input_text):
        orig_text_list = sent.split(" ")
        len_input = len(orig_text_list)
        if len_input > 3:
            rand_indx = random.randint(1, len_input-2)
            orig_text_list = orig_text_list[:rand_indx] + ["<mask>"] + orig_text_list[rand_indx:]
            new_text = " ".join(orig_text_list)
            augmented_text = unmasker_1(new_text)[0]['sequence']
            augmented_texts.append(augmented_text)
        else:
            augmented_texts.append(sent)
        
    text = ". ".join(augmented_texts)
    return text

from spacy.lang.en.stop_words import STOP_WORDS

def get_random_word_added_2(input_text):
    augmented_texts = []
    
    for sent in get_sents(input_text):
        orig_text_list = sent.split(" ")
        len_input = len(orig_text_list)
        if len_input > 3:
            rand_indx = random.randint(1, len_input-2)
            orig_text_list = orig_text_list[:rand_indx] + ["[MASK]"] + orig_text_list[rand_indx:]
            new_text = " ".join(orig_text_list)
            augmented_text = unmasker_2(new_text)[0]['sequence']
            augmented_texts.append(augmented_text)
        else:
            augmented_texts.append(sent)
        
    text = ". ".join(augmented_texts)
    return text

%%time 

train_data["augmented_text_1"] = train_data["reviewText"].apply(lambda x: get_random_word_added_1(x))

%%time 

train_data["augmented_text_2"] = train_data["reviewText"].apply(lambda x: get_random_word_added_2(x))

train_data

augmented_train_data_1 = train_data[["augmented_text_1", "overall", "sentiment"]]
augmented_train_data_1.rename(columns={"augmented_text_1": "reviewText"}, inplace=True)
augmented_train_data_2 = train_data[["augmented_text_2", "overall", "sentiment"]]
augmented_train_data_2.rename(columns={"augmented_text_2": "reviewText"}, inplace=True)
augmented_training_data = pd.concat([augmented_train_data_1, augmented_train_data_2], ignore_index=True)

augmented_training_data

augmented_training_data.to_csv("data_v2/Aug_RandomInsertion.csv", index=False)
