==> 04_DATA_AUGMENTATION_Validation_Data_Back_Translation.ipynb <==
import pandas as pd
from transformers import pipeline

val_data = pd.read_csv("data_v2/validation_data.csv")

en_to_fr = "Helsinki-NLP/opus-mt-en-fr"
en_to_it = "Helsinki-NLP/opus-mt-en-it"
en_to_de = "Helsinki-NLP/opus-mt-en-de"
en_to_es = "Helsinki-NLP/opus-mt-en-es"
en_to_zh = "Helsinki-NLP/opus-mt-en-zh"

fr_to_en = "Helsinki-NLP/opus-mt-fr-en"
it_to_en = "Helsinki-NLP/opus-mt-it-en"
de_to_en = "Helsinki-NLP/opus-mt-de-en"
es_to_en = "Helsinki-NLP/opus-mt-es-en"
zh_to_en = "Helsinki-NLP/opus-mt-zh-en"

tokenizer_kwargs = {
    "truncation": True, 
    "max_length": 400
}

translator_en_to_fr = pipeline("translation", model=en_to_fr, **tokenizer_kwargs)
translator_en_to_it = pipeline("translation", model=en_to_it, **tokenizer_kwargs)
translator_en_to_de = pipeline("translation", model=en_to_de, **tokenizer_kwargs)
translator_en_to_es = pipeline("translation", model=en_to_es, **tokenizer_kwargs)
translator_en_to_zh = pipeline("translation", model=en_to_zh, **tokenizer_kwargs)

translator_fr_to_en = pipeline("translation", model=fr_to_en, **tokenizer_kwargs)
translator_it_to_en = pipeline("translation", model=it_to_en, **tokenizer_kwargs)
translator_de_to_en = pipeline("translation", model=de_to_en, **tokenizer_kwargs)
translator_es_to_en = pipeline("translation", model=es_to_en, **tokenizer_kwargs)
translator_zh_to_en = pipeline("translation", model=zh_to_en, **tokenizer_kwargs)

reviewText_with_0 = val_data[val_data["sentiment"]==0]
reviewText = reviewText_with_0["reviewText"].tolist()

%%time 

french = translator_en_to_fr(reviewText)
french = [v["translation_text"] for v in french]
from_french = translator_fr_to_en(french)
from_french = [v["translation_text"] for v in from_french]



%%time 

italian = translator_en_to_it(reviewText)
italian = [v["translation_text"] for v in italian]
from_italian = translator_it_to_en(italian)
from_italian = [v["translation_text"] for v in from_italian]


%%time 

german = translator_en_to_de(reviewText)
german = [v["translation_text"] for v in german]
from_german = translator_de_to_en(german)
from_german = [v["translation_text"] for v in from_german]


%%time 

spanish = translator_en_to_es(reviewText)
spanish = [v["translation_text"] for v in spanish]
from_spanish = translator_es_to_en(spanish)
from_spanish = [v["translation_text"] for v in from_spanish]


%%time 

chinese = translator_en_to_zh(reviewText)
chinese = [v["translation_text"] for v in chinese]
from_chinese = translator_zh_to_en(chinese)
from_chinese = [v["translation_text"] for v in from_chinese]


reviewText_with_0["from_french"] = from_french
reviewText_with_0["from_italian"] = from_italian
reviewText_with_0["from_german"] = from_german
reviewText_with_0["from_spanish"] = from_spanish
reviewText_with_0["from_chinese"] = from_chinese

reviewText_with_from_french = reviewText_with_0[["from_french", "overall", "sentiment"]]
reviewText_with_from_italian = reviewText_with_0[["from_italian", "overall", "sentiment"]]
reviewText_with_from_german = reviewText_with_0[["from_german", "overall", "sentiment"]]
reviewText_with_from_spanish = reviewText_with_0[["from_spanish", "overall", "sentiment"]]
reviewText_with_from_chinese = reviewText_with_0[["from_chinese", "overall", "sentiment"]]

Aug_BackTranslation = pd.concat([reviewText_with_from_french, reviewText_with_from_italian, 
                                 reviewText_with_from_german, reviewText_with_from_spanish, 
                                 reviewText_with_from_chinese], ignore_index=True)

Aug_BackTranslation.to_csv("data_v2/Aug_Val_BackTranslation.csv", index=False)
